add, update, delete experience and education. Find Company By id and add pageable. Add filter for searching by keyword

+++ src/main/java/vn/hcmute/springboot/repository/CandidateExperienceRepository.java
@@ -3,6 +3,6 @@ package vn.hcmute.springboot.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.hcmute.springboot.model.CandidateExperience;
 
-public interface CandidateExperienceRepository extends JpaRepository<CandidateExperience, Long> {
+public interface CandidateExperienceRepository extends JpaRepository<CandidateExperience, Integer> {
 
 }
